refactor: change import of Alert component, test work of AppRoutes component
