+++ src/pages/active-applications-page/active-applications-page.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import styles from './active-applications-page.module.css';
 
-import { Alert } from '../../stories/alert/alert';
+import { Alert } from '../../ui/alert/alert';
 
 export function ActiveApplicationsPage() {
   return (
-    <main className={styles.main}>
+    <article className={styles.main}>
       <section className={styles.header_content}>
         <h1 className="text_type_main-extra-large">Заявки на участие</h1>
       </section>
@@ -13,6 +13,6 @@ export function ActiveApplicationsPage() {
         <Alert title="Активные заявки на участие отсутствуют" />
       </section>
       <section className={styles.footer_content}>Footer</section>
-    </main>
+    </article>
   );
 }
